+++ gui/main.py
@@ -1,7 +1,13 @@
 from PyQt5.QtCore import Qt
 from PyQt5.QtWidgets import (QApplication, QCheckBox, QGridLayout, QGroupBox,QComboBox,QLabel,QLineEdit,QMessageBox,QProgressBar,
-        QMenu, QPushButton, QRadioButton, QVBoxLayout,QHBoxLayout, QWidget)
+        QMenu, QPushButton, QRadioButton, QVBoxLayout,QHBoxLayout, QWidget,QBoxLayout)
+import os
+from encrypt import Encrypt
+from add import Add
+from delete import Delete
 
+filesystem=[]
+devicename=[]
 
 class Window(QWidget):
 	def __init__(self):
@@ -12,9 +18,9 @@ class Window(QWidget):
 		self.setWindowTitle("GUI")
 		self.setFixedSize(500,500)
 
-		self.Encrypt()
-		self.Add()
-		self.Delete()
+		Encrypt(self)
+		Add(self)
+		Delete(self)
 		self.Dcheck.toggle()		
 		self.Acheck.toggle()		
 		self.Echeck.toggle()
@@ -28,7 +34,7 @@ class Window(QWidget):
 			device=[]
 			
 			device1=[]
-			devicename=[]
+			
 			for line in f.readlines()[1:]:
 				device.append(line.split()[5])
 				#device2.append(line.split()[6:])
@@ -48,223 +54,7 @@ class Window(QWidget):
 					devicename.append(os.path.basename(s))
 		return devicename
 
-		
-
-
-		
-	def Encrypt(self):
-		self.Echeck=QCheckBox('Encrypt USB Drive',self)
-		
-		self.Echeck.stateChanged.connect(self.Entoggle)
-		
-		self.layout.addWidget(self.Echeck)
-
-				
-		self.EgroupBox = QGroupBox("")
-		self.EcomboBox = QComboBox()
-		self.EcomboBox.addItem(' Select your drive')
-		device=self.list_device()
-
-		for i in device:
-			self.comboBox.addItem(i)
-		
-		self.EcomboBox.setMinimumWidth(300)
-		   
-		
-		self.EcomboBox.move(150, 50)
-		
-		self.Ebtn=QPushButton("Format",self)
-		self.Ebtn.clicked.connect(self.format)
 	
-		
-		label=QLabel()
-		label.setText("Enter your Password")
-
-		self.Etextbox=QLineEdit()
-		self.Etextbox.setEnabled(False)
-		self.Etextbox.setEchoMode(QLineEdit.Password)
-		self.Etextbox.setMaxLength(10)
-		
-		label1=QLabel()
-		label1.setText("Confirm Your Password")
-
-		self.Etextbox1=QLineEdit()
-		self.Etextbox1.setEchoMode(QLineEdit.Password)
-		self.Etextbox1.setEnabled(False)
-		self.Etextbox1.setMaxLength(10)
-
-		self.Ebtn1=QPushButton("Finish",self)
-		self.Ebtn1.clicked.connect(self.Finish)
-		
-
-		vbox = QVBoxLayout()
-		
-		vbox.addWidget(self.EcomboBox)
-		vbox.addWidget(self.Ebtn)
-		vbox.addWidget(label)
-		vbox.addWidget(self.Etextbox)
-		vbox.addWidget(label1)
-		vbox.addWidget(self.Etextbox1)
-		vbox.addWidget(self.Ebtn1)
-		vbox.addStretch(1)
-		self.EgroupBox.setLayout(vbox)
-		
-		
-		self.layout.addWidget(self.EgroupBox)
-
-
-	def Add(self):
-		self.Acheck=QCheckBox('Add Keys',self)
-		self.Acheck.stateChanged.connect(self.Atoggle)
-		
-
-		self.layout.addWidget(self.Acheck)
-		
-		self.AgroupBox = QGroupBox("")
-		
-		self.AcomboBox = QComboBox()
-		self.AcomboBox.addItem(' Select your drive')
-		#device=self.list_device()
-
-		#for i in device:
-		#	self.comboBox.addItem(i)
-		
-		self.AcomboBox.setMinimumWidth(300)
-		 
-		
-		label=QLabel()
-		label.setText("Enter any existing password")
-
-		self.Atextbox=QLineEdit()
-		self.Atextbox.setEnabled(False)
-		self.Atextbox.setEchoMode(QLineEdit.Password)
-		self.Atextbox.setMaxLength(10)
-		
-		label1=QLabel()
-		label1.setText("Enter your new password")
-
-		self.Atextbox1=QLineEdit()
-		self.Atextbox1.setEchoMode(QLineEdit.Password)
-		self.Atextbox1.setEnabled(False)
-		self.Atextbox1.setMaxLength(10)
-
-		self.Abtn1=QPushButton("Finish",self)
-		#self.btn1.clicked.connect(self.Finish)
-		
-
-		vbox = QVBoxLayout()
-		
-		vbox.addWidget(self.AcomboBox)
-		vbox.addWidget(label)
-		vbox.addWidget(self.Atextbox)
-		vbox.addWidget(label1)
-		vbox.addWidget(self.Atextbox1)
-		vbox.addWidget(self.Abtn1)
-		vbox.addStretch(1)
-		self.AgroupBox.setLayout(vbox)
-		
-
-		self.layout.addWidget(self.AgroupBox)
-	
-
-
-	def Delete(self):
-		self.Dcheck=QCheckBox(' Delete Keys',self)
-		self.Dcheck.stateChanged.connect(self.Dtoggle)
-		
-
-		self.layout.addWidget(self.Dcheck)
-		
-		self.DgroupBox = QGroupBox("")
-		
-		self.DcomboBox = QComboBox()
-		self.DcomboBox.addItem(' Select your drive')
-		#device=self.list_device()
-
-		#for i in device:
-		#	self.comboBox.addItem(i)
-		
-		self.DcomboBox.setMinimumWidth(300)
-		 
-		
-		label=QLabel()
-		label.setText("Enter any existing password")
-
-		self.Dtextbox=QLineEdit()
-		self.Dtextbox.setEnabled(False)
-		self.Dtextbox.setEchoMode(QLineEdit.Password)
-		self.Dtextbox.setMaxLength(10)
-		
-		label1=QLabel()
-		label1.setText("Enter your new password")
-
-		self.Dtextbox1=QLineEdit()
-		self.Dtextbox1.setEchoMode(QLineEdit.Password)
-		self.Dtextbox1.setEnabled(False)
-		self.Dtextbox1.setMaxLength(10)
-
-		self.Dbtn1=QPushButton("Finish",self)
-		#self.btn1.clicked.connect(self.Finish)
-		
-
-		vbox = QVBoxLayout()
-		
-		vbox.addWidget(self.DcomboBox)
-		vbox.addWidget(label)
-		vbox.addWidget(self.Dtextbox)
-		vbox.addWidget(label1)
-		vbox.addWidget(self.Dtextbox1)
-		vbox.addWidget(self.Dbtn1)
-		vbox.addStretch(1)
-		self.DgroupBox.setLayout(vbox)
-		
-
-		self.layout.addWidget(self.DgroupBox)
-	
-
-
-
-
-	def Entoggle(self):
-		
-		
-		if self.Echeck.isChecked()==True:
-			self.EgroupBox.setVisible(True)
-			self.Acheck.setChecked(False)
-			self.Dcheck.setChecked(False) 			
-			
-			
-
-		else:
-			self.EgroupBox.setVisible(False)
-			#self.check1.setChecked(True)
-		
-	def Atoggle(self):
-		
-		
-		
-		if self.Acheck.isChecked()==True:
-			self.AgroupBox.setVisible(True)
-			self.Echeck.setChecked(False)
-			self.Dcheck.setChecked(False)
-		else:
-			
-			
-			self.AgroupBox.setVisible(False)
-			
-
-	def Dtoggle(self):
-		
-		
-		
-		if self.Dcheck.isChecked()==True:
-			self.DgroupBox.setVisible(True)
-			self.Echeck.setChecked(False)
-			self.Acheck.setChecked(False)
-		else:
-			
-			
-			self.DgroupBox.setVisible(False)
 			
 
 		
@@ -275,7 +65,7 @@ class Window(QWidget):
 			QMessageBox.information(self,'Alert',"Please Select Drive",QMessageBox.Close)
 		else:
 
-			choices=QMessageBox.question(self,'Message',"To format disk you need to enter password.Do you want to continue?",QMessageBox.Yes | QMessageBox.No,QMessageBox.No)
+			choices=QMessageBox.question(self,'Message',"To format disk you need to enter user password.Do you want to continue?",QMessageBox.Yes | QMessageBox.No,QMessageBox.No)
 			self.complete=0
 		
 			if choices==QMessageBox.Yes:
@@ -298,7 +88,7 @@ class Window(QWidget):
 				self.widget1=QWidget()
 				self.widget1.setLayout(self.hboxpass)
 				self.widget1.setWindowTitle('Password window')
-				self.widget1.setGeometry(50,50,500,100)
+				#self.widget1.setGeometry(50,50,500,100)
 				self.widget1.show()
 
 				btncontinue.clicked.connect(self.passwordfun)
@@ -310,7 +100,7 @@ class Window(QWidget):
 	def passwordfun(self):
 			
 			
-			self.widget1.close()
+			#self.widget1.close()
 			self.progressLabel = QLabel('Progress Bar:', self)
 		
 			# Creating a progress bar and setting the value limits
@@ -328,29 +118,39 @@ class Window(QWidget):
 			self.hboxLayout.addWidget(self.progressBar)
 			self.hboxLayout.addWidget(self.btn)
 			
-			command='wipefs -a --force %s'%filesystem[devicename.index(self.EcomboBox.currentText())]
+			command='umount %s'%filesystem[devicename.index(self.EcomboBox.currentText())]
+			command1='wipefs -a %s'%filesystem[devicename.index(self.EcomboBox.currentText())]
 			
-			x=os.system("echo %s |sudo -S %s"%(self.password.text(),command))			
+			x=os.system("echo %s |sudo -S %s"%(self.password.text(),command))
+			
 			if x==0:
-				# Setting the hBoxLayout as the main layout
-				self.widget=QWidget()
-				self.widget.setLayout(self.hboxLayout)
-				self.widget.setWindowTitle('Dialog with Progressbar')
-				self.widget.setGeometry(50,50,500,100)
-				self.widget.show()
+				y=os.system("echo %s |sudo -S %s"%(self.password.text(),command1))			
+				if y==0:
+					# Setting the hBoxLayout as the main layout
+					self.widget=QWidget()
+					self.widget.setLayout(self.hboxLayout)
+					self.widget.setWindowTitle('Dialog with Progressbar')
+					self.widget.setGeometry(50,50,500,100)
+					self.widget1.close()
+					self.widget.show()
 
 			
 		
-				connect=0
-				while connect<100:
-					connect+=0.001				
-					self.progressBar.setValue(connect)
+					connect=0
+					while connect<100:
+						connect+=0.001				
+						self.progressBar.setValue(connect)
 			
-				self.widget.close()
-				QMessageBox.information(self,'Message',"Drive is formatted.Now you can set password",QMessageBox.Ok) 
-				self.Etextbox.setEnabled(True)
-				self.Etextbox1.setEnabled(True)
+					self.widget.close()
+					QMessageBox.information(self,'Message',"Drive is formatted.Now you can set password",QMessageBox.Ok) 
+					self.Etextbox.setEnabled(True)
+					self.Etextbox1.setEnabled(True)
+				else:
+					choice=QMessageBox.information(self,'Error',"Cannot format drive.Please verify that password is correct ",QMessageBox.Retry|QMessageBox.Close ,QMessageBox.Retry)
 
+					if choice==QMessageBox.Retry:
+						self.widget1.show()
+				
 			
 			
 				
@@ -383,4 +183,5 @@ if __name__ == '__main__':
     app = QApplication(sys.argv)
     window = Window()
     window.show()
+
 sys.exit(app.exec_())
